# POC new memory schedule algo